requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2==2.9.6
sqlparse==0.4.2
whitenoise==6.2.0 
certifi==2022.12.7
charset-normalizer==3.1.0
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
requests==2.28.2
tzdata==2022.7
urllib3==1.26.15